% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example1.R
\docType{data}
\name{example1}
\alias{example1}
\title{Missouri Dickcissel Nest Survival Data - Interval Format}
\format{A data frame with 217 observations on the following 8 variables.:
  \describe{
    \item{\code{id}}{Nest id}
    \item{\code{fate}}{Nest fate; if 1, the nest fate is success;if 0, the
    nest fate is failure.}
    \item{\code{ul}}{The youngest possible age that the nest could have been
    when first encounterd.}
    \item{\code{ur}}{The oldest possible age that the nest could have been
    when first encounterd.}
    \item{\code{zl}}{The smallest possible number of time units from the
    first encounter date to the outcome date.}
    \item{\code{zr}}{The largest possible number of time units from the first
     encounter date to the outcome date.}
    \item{\code{cov1}}{The first nest-specific covariate.}
    \item{\code{cov2}}{The second nest-specific covariate.}
  }}
\usage{
example1
}
\description{
The Missouri dickcissel dataset collected by Suedkamp Wells(2005)
is analyzed via BEANSP for illustration. For more detailed analysis, see Cao et.al(2009).
}
\references{
Cao, J., He, C., Suedkamp Wells, K.M., Millspaugh, J.J.,
and Ryan, M.R. (2009).  Modeling age and nest-specific survival using a
hierarchical Bayesian approach. Biometrics, 65, 1052-1062.
}
\keyword{datasets}
