\name{plot.asr}
\alias{plot.asr}
\title{
Plot average survival rate 
}
\description{
This is nest-average age-specific survival rate plot.
}
\usage{
\method{plot}{asr}(x, jj, xlab = "nest age", 
            ylab = "average age-specific nest survival rate", ...)
}
\arguments{
  \item{x}{
average survival rate. By default, x = out$asr.
}
  \item{jj}{
nest period time
}
  \item{xlab}{
label x as nest age
}
  \item{ylab}{
label y as average age-specific nest survival rate
}
  \item{\dots}{
plot
}
}
\keyword{ methods }
