\name{plot.del}
\alias{plot.del}
\title{
Plot age-specific encounter rate
}
\description{
This is age-specific encounter rate plot.
}
\usage{
\method{plot}{del}(x, jj, xlab = "nest age",
            ylab = "individual age-specific nest encounter rate", ...)
}
\arguments{
  \item{x}{
age-specific nest encoutner rate.By default,x = out$del.
}
  \item{jj}{
nest period time
}
  \item{xlab}{
label x as nest age
}
  \item{ylab}{
label y as age-specific nest encoutner rate
}
  \item{\dots}{
plot
}
}
\keyword{ methods }
