\name{plot.sr}
\alias{plot.sr}
\title{
Plot age-specific nest survival rate.
}
\description{
This is age-specific survival plot.
}
\usage{
\method{plot}{sr}(x, n, jj, xlab = "nest age", 
               ylab = "individual age-specific nest survival rate", ...)
}
\arguments{
  \item{x}{
age-specific survival rate. By default,x = sr.
}
  \item{n}{
the index of nest. For example, n=1 is the 1st nest.
}
  \item{jj}{
nest period time
}
  \item{xlab}{
label x as nest age
}
  \item{ylab}{
label y as age-specific survival rate
}
  \item{\dots}{
plot
}
}
\keyword{ methods }
