% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_asr.R
\name{plot_asr}
\alias{plot_asr}
\title{Plot Average Survival Rate}
\usage{
plot_asr(x, xlab = "nest age",
  ylab = "average age-specific nest survival rate", ...)
}
\arguments{
\item{x}{a \code{nsr} object}

\item{xlab}{label of \code{x}-axis}

\item{ylab}{label of \code{y}-axis}

\item{...}{arguments to be passed to methods.}
}
\description{
This is nest-average age-specific survival rate plot.
}
\examples{

 n0 <- as.integer(10)
 ntotal <- as.integer(110)
 a <- as.double(rep(2.0,2))
 b <- as.double(rep(1.0,2))
 sigma <- as.double(rep(1.0,2))

 data("example1")
 jj <- as.integer(19)
 nx <- as.integer(2)
 nn <- as.integer(217)
 x1 <- example1[1:nn,7]
 x2 <- example1[1:nn,8]
 x <- cbind(as.double(x1),as.double(x2))
 temp <- nestconv(0,nn,jj,example1, censored  =  FALSE)
 ul <- as.integer(temp[,2])
 ur <- as.integer(temp[,3])
 zl <- as.integer(temp[,4])
 zr <- as.integer(temp[,5])
 y <- as.integer(temp[,6])

 day <- as.double(rep(0.0,jj))
 enc <- as.double(rep(0.0,jj - 1))
 covar <- as.double(rep(0.0,nx))
 out <- nestsr(jj = jj, nx = nx, nn = nn, zl = zl, zr = zr, x = x, y = y,
               a = a, b = b, sigma = sigma, day = day, enc = enc, covar = covar,
               n0 = n0, ntotal = ntotal, ul = ul, ur = ur)
 plot_asr(x=out)
}
\author{
Chong He, Yiqun Yang, Jing Cao, Peng Shao
}
\keyword{methods}
