% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_covar_trace.R
\name{plot_covar_trace}
\alias{plot_covar_trace}
\title{Trace Plot for Covariate Effect}
\usage{
plot_covar_trace(x, j, n0, ntotal, xlab = "iter",
  ylab = paste("encounter age effect at j=", j), ...)
}
\arguments{
\item{x}{a \code{nsr} object}

\item{j}{The j-th covariate effect.j=1,2,...,jj-1}

\item{n0}{The number of burn-in cycles.}

\item{ntotal}{The number of total Gibbs cycles.}

\item{xlab}{label of \code{x}-axis}

\item{ylab}{label of \code{y}-axis}

\item{...}{Arguments to be passed to methods.}
}
\description{
Displays a plot of sampled values for the covariate effect
vs. iterations.
}
\author{
Chong He, Yiqun Yang, Jing Cao, Peng Shao
}
\keyword{methods}
