% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_enc_trace.R
\name{plot_enc_trace}
\alias{plot_enc_trace}
\title{Trace plot for enc}
\usage{
plot_enc_trace(x, j, n0, ntotal, xlab = "iter",
  ylab = paste("encounter age effect at j=", j), ...)
}
\arguments{
\item{x}{a \code{nsr} object}

\item{j}{The j-th age effect.j=1,2,...,jj-1}

\item{n0}{The number of burn-in cycles.}

\item{ntotal}{The number of total Gibbs cycles.}

\item{...}{Arguments to be passed to methods.}
}
\description{
Displays a plot of iterations vs. sampled values for the age
effect of encounter rate.
}
\author{
Chong He, Yiqun Yang, Jing Cao, Peng Shao
}
\keyword{methods}
