% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.nsr.R
\name{summary.nsr}
\alias{summary.nsr}
\alias{print.summary.nsr}
\alias{print.summary.nsr}
\title{Summary \code{nsr} Object}
\usage{
\method{summary}{nsr}(object, covnames = FALSE, ...)

\method{print}{summary.nsr}(x, ...)
}
\arguments{
\item{object}{An object of class "nsr".}

\item{covnames}{A character vector of the names of covariates.}

\item{...}{additional arguments affecting the summary produced.}

\item{x}{An object of class "summary.nsr".}
}
\value{
An object of class "summary.nsr" containing the following components:
\describe{
\item{\code{covar.table}}{A data frame of mean, standard deviation, 2.5% quantile
and 97.5 quantile of covariate(s).}
\item{\code{DIC}}{Model selection criterion DIC value}
\item{\code{WAIC}}{Model selection criterion WAIC value}
\item{\code{n0}}{The number of burn-in cycles.}
\item{\code{ntotal}}{The number of total Gibbs cycles.}
}
}
\description{
Convert date format data or interval format data to the standard format.
}
\examples{

data("example4")
n0 <- as.integer(10)
ntotal <- as.integer(110)
a <- as.double(rep(2.0,2))
b <- as.double(rep(1.0,2))
sigma <- as.double(rep(1.0,2))
jj <- as.integer(26)
nx <- as.integer(1)
nn <- as.integer(221)
x <- as.double(as.numeric(example4[,7]) - 1)
temp <- nestconv(1,nn,jj,example4, censored  =  TRUE)
zl <- as.integer(temp[,2])
zr <- as.integer(temp[,3])
y <- as.integer(temp[,4])
mi <- as.integer(example4[,6])

day <- as.double(rep(0.0,jj))
enc <- as.double(rep(0.0,jj - 1))
covar <- as.double(rep(0.0,nx))
tha <- as.double(rep(0.0,2))
mutha <- as.double(rep(0.0,2))
vartha <- as.double(rep(5.0,2))
out <- nestsr(jj = jj, nx = nx, nn = nn, zl = zl, zr = zr, x = x, y = y, a = a,
              b = b, sigma = sigma, day = day, enc = enc, covar = covar, n0 = n0,
              ntotal = ntotal, mi = mi, tha = tha, mutha = mutha, vartha = vartha)
summary(out)

data("example1")
jj <- as.integer(19)
nx <- as.integer(2)
nn <- as.integer(217)
x1 <- example1[1:nn,7]
x2 <- example1[1:nn,8]
x <- cbind(as.double(x1),as.double(x2))
temp <- nestconv(0,nn,jj,example1, censored  =  FALSE)
ul <- as.integer(temp[,2])
ur <- as.integer(temp[,3])
zl <- as.integer(temp[,4])
zr <- as.integer(temp[,5])
y <- as.integer(temp[,6])

day <- as.double(rep(0.0,jj))
enc <- as.double(rep(0.0,jj - 1))
covar <- as.double(rep(0.0,nx))
out <- nestsr(jj = jj, nx = nx, nn = nn, zl = zl, zr = zr, x = x, y = y,
              a = a, b = b, sigma = sigma, day = day, enc = enc, covar = covar,
              n0 = n0, ntotal = ntotal, ul = ul, ur = ur)
summary(out)

}
\author{
Chong He, Yiqun Yang, Jing Cao, Peng Shao
}
\keyword{methods}
