% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example3.R
\docType{data}
\name{example3}
\alias{example3}
\title{National Mourning Dove Nest Survival Data - Interval Format}
\format{A data frame with 217 observations on the following 6 variables.:
  \describe{
    \item{\code{id}}{Nest id}
    \item{\code{zl}}{The smallest possible number of time units from the
    first encounter date to the outcome date.}
    \item{\code{zr}}{The largest possible number of time units from the first
     encounter date to the outcome date.}
    \item{\code{Fate}}{Nest fate; if 1, the nest fate is success;if 0, the
    nest fate is failure.}
    \item{\code{Censor}}{The missing fate indicator; if 1, the nest fage is missing;
    if 0, the nest fate is not missing}
    \item{\code{cov1}}{The first nest-specific covariate.}
  }}
\usage{
example3
}
\description{
The original data were collected by Geissler et al.(1987). This
example dataset only contains a fragment of whold dataset.
}
\references{
blank
}
\keyword{datasets}
