% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestconv.R
\name{nestconv}
\alias{nestconv}
\title{A test}
\usage{
nestconv(type, nn, jj, data, censored = TRUE)
}
\arguments{
\item{x}{this is a number}
}
\value{
a

b
}
\description{
a test function.
}
\references{
Cao, J., He, C., Suedkamp Wells, K.M., Millspaugh, J.J., and
Ryan, M.R. (2009).  Modeling age and nest-specific survival using a hierarchical
Bayesian approach. Biometrics, 65, 1052-1062.
}
\author{
Chong He, Yiqun Yang, Jing Cao, Peng Shao
}
\keyword{methods}
