% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestconv.R
\name{nestconv}
\alias{nestconv}
\title{Data Format Conversion}
\usage{
nestconv(type, nn, jj, data, censored = TRUE)
}
\arguments{
\item{type}{If data are interval format, type=0; if data are date format, type=1.}

\item{nn}{The number of observed nests (sample size).}

\item{jj}{The number of time units(days) a nest is required to survive to be considered successful.}

\item{data}{Input data.}

\item{censored}{If data are censored, censored=TRUE; if data are not censored, censored=FALSE}
}
\value{
The returned values are \code{zl,zr}. If censored=FALSE, \code{ul,ur} are also returned.
}
\description{
Convert date format data or interval format data to the standard format.
}
\examples{

data("example1")
jj <- as.integer(19)
nn <- as.integer(217)
temp1 <- nestconv(0,nn,jj,example1, censored  =  FALSE)
data("example2")
jj <- as.integer(20)
nn <- as.integer(233)
temp2 <- nestconv(1,nn,jj,example2, censored  =  FALSE)
data("example3")
jj <- as.integer(26)
nn <- as.integer(221)
temp3 <- nestconv(0,nn,jj,example3, censored  =  TRUE)
data("example4")
jj <- as.integer(26)
nn <- as.integer(221)
temp4 <- nestconv(1,nn,jj,example4, censored  =  TRUE)

}
\references{
Cao, J., He, C., Suedkamp Wells, K.M., Millspaugh, J.J., and
Ryan, M.R. (2009).  Modeling age and nest-specific survival using a hierarchical
Bayesian approach. Biometrics, 65, 1052-1062.
}
\author{
Chong He, Yiqun Yang, Jing Cao, Peng Shao
}
\keyword{methods}
